+++ src/main/java/com/example/calculator/service/CalculatorServicee.java
@@ -3,11 +3,11 @@ package com.example.calculator.service;
 public interface CalculatorServicee {
     String hello();
 
-    int addTwoNum(int num1, int num2);
+    long addTwoNum(long num1, long num2);
 
-    int minusTwoNum(int num1, int num2);
+    long minusTwoNum(long num1, long num2);
 
-    int multiplyTwoNum(int num1, int num2);
+    long multiplyTwoNum(long num1, long num2);
 
     double divideTwoNum(int num1, int num2);
 }
